Pagina web Casí terminada

+++ AppWebBD/Models/EstadoCuenta.cs
@@ -5,7 +5,7 @@ namespace AppWebBD.Models
     public class EstadoCuenta
     {
         public int id { get; set; }
-        public long NumeroCuenta { get; set; }
+        public int NumeroCuenta { get; set; }
         public string FechaInicio { get; set; }
         public string FechaFin { get; set; }
         public long SaldoInicial { get; set; }
